Use css ::before pseudo element for ad labels
